ci(Update): Refactor map click handler and add edit feature
The handling of map clicks has been refactored into a separate function,
`SetClickTPlace`. This simplifies the main event handler and provides
more clarity. In addition, an edit feature has been added to the context
menu which allows users to edit a clicked map element by bringing up an
editing interface. Display of this edit feature in the context menu is
dynamically controlled based on whether or not the selected map feature
is editable.