ci(Update): Change database path in OnConfiguring
The database path in the OnConfiguring method has been updated to
reflect the correct location of the database in the DataBaseContext.cs
file. The previous database path, which was commented out, is now eased
back into use, while the recent actively used path has been commented
out for future reference.