ci(Update): Add Mapsui layers and feature to WindowEdit
Three new import statements have been added (Mapsui, Mapsui.Layers, and
Mapsui.Styles) to support the addition of new layers and features to the
map in WindowEdit.xaml.cs. The changes improve the WindowEdit
constructor's map setup and the SetTplace function's marking
functionality. Now, features can be added to the WritableLayer, and the
map will automatically center and zoom on these features.