ci(Update): Add temporary marker feature to MapControl
A new feature has been added that enables temporary marker positioning
in the MapControl component. An Info event handler, MapControl_OnInfo,
is included and triggers with a tap on the map, placing a marker at the
selected location. If there's already a temporary marker, the old one is
removed before the new one is positioned.