ci(Update): Rename Nominatim related classes for clarity
The classes `NominatimStruc` and `NominatimAddress` in the
`MyExpenses.Models.WebApi.Nominatim` namespace have been renamed to
`NominatimSearchResult` and `NominatimDetailedAddress` respectively.
This provides a clearer context for their usage and the information they
represent. The changes also extend to related variable and method return
types where these classes are referred to.