ci(Update): Update WindowEdit to enhance map interactions
The changes include the addition of a clear flag to the SetTplace
function to optionally clear the WritableLayer. Clearing is performed
with the new code. The HandleNominatimResult function is also updated to
handle the mapping of a searched place to TPlace type and setting it
into the map, providing a better response to the user search action.