ci(Update): Implement zoom functionality in WindowEdit
The code adds a zoom to point functionality in the WindowEdit file. It
includes edge cases to handle when there's more than one point, allowing
for a zoomed-in view on a group of points, and when there's only one
point, the function will center on that point and zoom to it. A margin
percentage is also added to set a buffer zone around the zoomed area.