ci(Update): Add ZoomToMPoint method and update ButtonValidNewPoint logic
The ZoomToMPoint method was added to centralize the logic for zooming to
a specific point. As part of this change, the logic in the
ButtonValidNewPoint_OnClick method was updated. This includes validation
and point removal, as well as the addition of the new method to zoom
into the added point.