ci(Update): Implement AutoMapper for PointFeature conversion
The current changes involve the implementation of AutoMapper for
converting and mapping data between the TPlace and PointFeature objects
in the MyExpenses application. Previously, the ToPointFeature method was
used for this, but the new change replaces this with an AutoMapper,
increasing the application's flexibility and maintainability. Moreover,
updates to PointFeature handling in MainWindow.xaml.cs,
WindowEdit.xaml.cs, and WindowSelectNominatimSearchResult.xaml.cs files
were made to accommodate and utilize this new mapping approach.