ci(Update): Implement feature deletion functionality
The commit provides the implementation of the feature deletion
functionality that was previously marked as TODO. It adds the logic to
remove a feature from the database and the WritableLayer. If the
operation is successful, a message box informs the user; if an exception
occurs, an error message is shown.