ci(Update): Add total by accounts to MainWindow
In MainWindow.xaml.cs, a new property VTotalByAccounts was added to
store a list of total values by account from DataBaseContext. This list
is now being utilized in MainWindow.xaml within an ItemsControl element,
replacing the hard-coded buttons for individual accounts with dynamic
buttons that are bound to the data from VTotalByAccounts.