ci(Update): Update Total property accessibility and MainWindow UI
The accessibility of the 'Total' property in MainWindow.xaml.cs has been
changed from private to public. Also, a new StackPanel with a Label and
materialDesign:PackIcon was added to MainWindow.xaml to present 'Total'
value interactively with visual cues. The XAML namespaces were updated
to accommodate these changes, including the addition of converters and
materialDesign.