ci(Update): Update Dashboard to show current month expenses
In this commit, the UpdateGraph method in the Dashboard page has been
updated to only display the expenses for the current month. A new
variable was added to get the current date and this was used as a filter
when calculating category totals. Additionally, pieSeries tooltip now
shows amount in local currency.