ci(Update): Refactor decimal validation into DecimalRegex class
The commit removes the in-line regular expression for validating decimal
input from the TextFieldStartingBalance_OnPreviewTextInput method in
AddAccount.xaml.cs. The validation logic is now refactored into a new
static extension method, IsOnlyDecimal, in the DecimalRegex class in the
MyExpenses.Wpf.Ressources.Regex namespace. The purpose of this
refactoring is to improve code maintainability and reusability.