ci(Update): Add validation for account creation
In the dashboard and add account pages, code was added to return from
the add account dialog if the dialog result is not true. This means that
account creation doesn't proceed unless it's properly validated first.
Plus, the "Valid" button in the account creation dialog now sets
DialogResult to true, indicating successful validation.