ci(Update): Add logging capabilities using Serilog
Introduced a new utility project which includes the configuration for
the Serilog logger. Added references to this new utility project in the
main application project. Enhanced the startup process of the
application to instantiate the logger and manage its lifetime, notably
at application's startup and shutdown, as well as during unhandled
exceptions.