ci(Update): Refine error message when account name is empty
The error message for an empty account name has been made more specific
and accurate. Changes were made in the AddAccount page and in the
resource files for the application, including language-specific resource
files. The aim is to provide clearer feedback to the user when an
account name is not provided.