ci(Update): Implement localization for error message
The error message which indicates that the starting balance description
cannot be empty, is now localized. This change is reflected in the
AddAccount.xaml.cs, AddAccountResources.Designer.cs,
AddAccountResources.resx, AddAccountResources.fr-fr.resx, and
AddAccountResources.en-001.resx files. This increases accessibility and
ensures clearer communication with users in different locales.