ci(Update): Implement logging for adding new account type
Serilog is now used to log the attempt to add a new account type to the
system. Log messages were added to indicate either the successful
addition of this new type or any possible errors during this process.
Using this logging library should enhance the diagnostic potential of
our application.