ci(Update): Remove delete button from account type and currency
The delete button present in the account type and currency sections has
been removed. Changes apply to various language resource files where the
button title was declared and used. The corresponding code in XAML and
C# where the button was used and expected to be present has also been
adjusted.