ci(Update): Refactor and add RefreshAccountTotal method
A new method named RefreshAccountTotal was added to update
VTotalByAccounts property in the DashBoard class. Also, the type of
VTotalByAccounts property was changed from List to ObservableCollection,
which will now reflect changes in UI in real-time. Lastly, the
RefreshAccountTotal method was called in DashBoard's constructor and
after adding an account to refresh the data immediately.