ci(Update): Change VHistories type to ObservableCollection and add RefreshDataGrid
The type of VHistories property in DashBoard.xaml.cs has been changed
from List to ObservableCollection. Also, a new method RefreshDataGrid
has been added which is responsible for refreshing the DataGrid,
factoring in the account name selected in order to display the
appropriate data. The filling of the VHistories collection has been
moved from the DashBoard constructor to this method.