ci(Update): Updated LocationManagementPage layout and added selection functionality
Added a ComboBox to the LocationManagementPage.xaml for KnownTileSources
selection capability. Also updated the layout of the page to improve
usability, including setting a minimum width for the first column and
adding a dynamic background to the MapControl. Addition of the
corresponding selection changed function in the code-behind to manage
selected value.