ci(Update): Add SplitUpperCaseWordsConverter and integrate with LocationManagementPage
This commit includes the addition of the `SplitUpperCaseWordsConverter`
and its integration in the `LocationManagementPage`. The converter is
used within the ComboBox in the XAML file to format the displayed text,
making it more readable by splitting upper case words. Furthermore, a
new regex method `SplitUpperCaseWords` is implemented in `StringRegex`
to support this feature.