ci(Update): Add initial zoom feature for locations in Map Control
In the file 'LocationManagementPage.xaml.cs', a new method
'SetInitialZoom()' has been added. This method sets the initial zoom
level based on the number of available points on the map. For multiple
points, it calculates the min and max coordinates to create a bounding
rectangle, then zooms the map to fit that rectangle. Adjustments for
margins are also added.