ci(Update): Updated application icon and its placement in the title bar
In MainWindow.xaml, the default option to show the icon in the title bar
was disabled in favor of manually placing it on the left side of the
window. Additionally, icon resources were added to the WPF project and
will be copied to the output directory during the build process.