ci(Update): Add cascade delete functionality in DbContextHelper
A cascade option has been added for deletion in DbContextHelper, which
triggers the removal of child entities when a parent entity is deleted.
New reflection methods have been used to retrieve the navigation
properties of the entity to find and delete all related children. This
functionality is optional and can be turned off by setting the cascade
parameter to false.