ci(Update): Add currency symbol to TotalStr in DashBoardPage
The DashBoardPage.xaml.cs file has been updated to include the currency
symbol in the 'TotalStr' property. This was accomplished by adding a new
private string property 'Symbol', which notifies 'TotalStr' of any
changes to its value. The symbol is also read from the 'VTotalByAccount'
object whenever a button is clicked.