ci(Update): Add account management components to WPF page
The AccountManagementPage in the WPF app now includes a binding to
TotalByAccounts. This comes with an ItemsControl element that displays
information about each account, including a dynamically converted icon
based on the account's total and additional account details like Name
and Symbol. Necessary converters and namespaces were also added as part
of this update.