ci(Update): Add click action for account button in AccountManagementPage
The functionality for the click event of the account button in the
AccountManagementPage has been added. This new implementation fetches
the account data context on button click and sends it to the console.
The XAML markup has also been updated to link the button with its
corresponding click event handler.