ci(Update): Added SQL converter utility and implemented it in AccountManagementPage
The changes include the addition of a new utility in the Utils project,
specifically for SQL conversions. In addition to adding the new SQL
converter utility, it's also been implemented in the
AccountManagementPage. This refactoring has shifted the responsibility
for retrieving view-model data from the view layer into the utility
function.