ci(Update): Add delete functionality to `AddEditAccountWindow`
Two new functionalities have been added to the `AddEditAccountWindow` to
enhance the user experience. Firstly, a `SetTAccount` method has been
implemented which copies the properties of the 'account' to the
'Account' field and also sets `EditAccount` to true. Also, a
`ButtonDelete_OnClick` event handler has been added to handle delete
request on click of the button. Visibility of the "Delete" button is
also managed based on the `EditAccount` field. Changes are made in the
`AccountManagementPage.xaml.cs` to call `SetTAccount` method when an
account is selected.