ci(Update): Rename 'compte_fk' field to 'account_fk' in database and models
The field 'compte_fk' was renamed to 'account_fk' in various parts of
the code comprising the MyExpenses application. These changes are
reflected in the THistory model, 'Definition.sql', and 'command.sql'
files. The schema and code now align better with the application
language, making it more intuitive for an English-speaking development
team.