ci(Update): Add dynamic button visibility in message box
In the message box class, the ability to dynamically set button
visibility based on the MessageBoxResult was introduced. Now, depending
on the MessageBoxResult, different combinations of 'Yes', 'No', 'Ok',
and 'Cancel' buttons will be visible. The properties were also
reorganized for cleaner code. In the XAML file, names were added to the
buttons for this functionality.