ci(Update): Add localization to MsgBox buttons in MyExpenses app
A new set of resources has been added to the MyExpenses.Wpf project for
localization of the message box buttons. Accordingly, the message box
buttons' contents have been linked to new properties in the
MsgBoxMessageWindow.xaml.cs file. The changes allow for the button's
content to be translated into different languages as defined in the
resource files.