ci(Update): Implement functionality for 'Add Category Type' button
The 'Add Category Type' button on the 'CategoryTypeManagement' page now
opens a window for adding a new category type. If a new type is added
successfully, it will be sorted and added to the existing list of types
and the user will be notified. Errors during this process are logged and
also presented to the user.