ci(Update): Refactor property name and type in CategoryTypeManagementPage
The name of the property "EditCategoryTypeOriginalName" in the
AddEditCategoryTypeWindow class has been changed to "EditCategoryType"
and the type has been changed from string to bool. This change was
necessitated to better reflect the property's purpose and usage within
the class. The CategoryTypeManagementPage class has been updated
accordingly to reflect this change.