ci(Update): Add DeepCopy method to TCategoryType model
We've added a DeepCopy method to the TCategoryType model in the SQL
Tables section. This method allows for an easier and more concise way to
create a deep copy of an object. Consequently, the
CategoryTypeManagementPage.xaml.cs file has been updated to use this new
function instead of manually copying each property.