ci(Update): Add color preview and converter to ColorPickerControl
A ColorToHexConverter has been added to the resources of the
ColorPickerControl.xaml in the UserControl. The order of the Grid column
definitions has been altered. A Label and Rectangle element for color
preview have also been added to the Grid. These changes allow the
application to display the color selected in hex format.