ci(Update): Add "can_be_deleted" column to t_place table
A new column "can_be_deleted" is added to the t_place table in the
Definition.sql, used to determine if a place can be deleted or not. This
change is reflected in the model TPlace.cs as a new property. A default
value for "can_be_deleted" is set in the data context.