ci(Update): Refactor database schema and associated models
This commit addresses the simplification and refactoring of the
application's database schema, as well as the corresponding models.
Changes include the removal of several unneeded relationships in the
THistory, TBankTransfer, and TAccount models. Besides, a new
relationship 'BankTransferFk' was added in 'THistory' model, and the
'TBankTransfer' model now implements the 'ISql' interface. The
'VHistory' view was also updated to include the 'main_reason' column.