ci(Update): Add 'ToAccount' related functionality and properties in BankTransferPage
Added new 'VToAccount' and 'VToAccountIncrease' dependency properties
for managing account transfer scenarios in the BankTransferPage. This
update also includes the logic to refresh the lists of 'ToAccounts' and
'FromAccounts' based on selected accounts. This allows for a more
dynamic and efficient account selection process in the UI. Changes also
include the handling of balance updates based on the performed
transactions.