ci(Update): Add hint text to form fields in BankTransferPage
For better visibility and user understanding, descriptive hint text has
been added to the form fields on the BankTransferPage. Every ComboBox
and TextBox now provide correlated hints, such as 'From account :' and
'To account :'. A slight modification has also been made to the
DatePicker hint to follow the same pattern.