ci(Update): Add new text fields to BankTransferPage
Two additional read-only text fields were added to the BankTransferPage.
These fields are for displaying the 'MainReason' and 'AdditionalReason'
properties of a BankTransfer, enhancing the detail and context provided
to the user. Modified the row setup of the Grid and the location of the
existing StackPanel to accommodate these changes.