ci(Update): Added account editing functionality to BankTransferPage
The function `ButtonFromAddAccount_OnClick` in
`BankTransferPage.xaml.cs` has been updated to include account editing
functionality through the `AddEditAccountWindow`. If the user chooses to
delete an account, it is removed from the respective lists. Accounts can
now be added or edited from this method, and errors are logged if the
addition or modification is unsuccessful.