ci(Update): Refactor account removal in BankTransferPage
The method for removing accounts in the BankTransferPage.xaml.cs file
was refactored. The account Id used for comparison in the removal of
accounts in various lists (Accounts, FromAccounts, ToAccounts) was
changed from BankTransfer.FromAccountFk and BankTransfer.ToAccountFk to
a single variable, accountId. This makes the code more maintainable and
easier to understand.