ci(Update): Added textbox and button for recording expense value
A textbox and a button were added to the RecordExpensePage.xaml file.
The textbox allows the user to input the amount of an expense, while the
button will trigger a calculator feature. The related UI interactions
and value changes were also added in the corresponding C# file
(RecordExpensePage.xaml.cs).