ci(Update): Add new checkbox to RecordExpensePage
A new checkbox component has been added to the RecordExpensePage in both
the XAML markup and the corresponding C# file. This checkbox, "Is
pointed", allows the user to mark if the expense is pointed or not. Some
properties like HorizontalContentAlignment, VerticalContentAlignment,
etc., have been set to align the checkbox in a user-friendly manner.