ci(Update): Add MapControl and tile selector to RecordExpensePage.xaml
Implemented a map control on the RecordExpensePage, and added a ComboBox
for users to select different tile sources. The ComboBox utilizes a new
SplitUpperCaseWordsConverter to improve readability of the list items.
The map control is filled with predefined KnownTileSources, and enables
the selection and visual inspection of different map tile layers.