ci(Update): Enhanced payment mode management in RecordExpensePage
Added check to see if a payment mode can be changed or deleted. If a
mode isn't editable, an error message is displayed to the user. Further
improvements include logic for editing the payment mode and handling the
outcomes successfully or unsuccessully, providing relevant user feedback
for both scenarios. Also implemented logs for tracking the functionality
of mode payment.