ci(Update): Refactor UpdateMapPoint to handle null places
Moved the null check for places from SelectorPlace_OnSelectionChanged to
inside UpdateMapPoint method. Now, the UpdateMapPoint method can handle
null places, which triggers a clear and refresh of the MapControl. If
the place is not null, it proceeds to convert the place into a feature,
adds it to the PlaceLayer, and adjusts the MapControl accordingly.