ci(Update): Implement tile source selection and tile layer update
The update enables users to select a tile source from a ComboBox. The
selected tile source is used to create a new tile layer which is then
added to the Map Control in the AddEditLocationWindow. This update also
includes some code cleanup in the AddEditLocationWindow.xaml and
AddEditLocationWindow.xaml.cs files.