ci(Update): Refactor RecordExpensePage for an editable place
Removed unused piece of code and implemented a failure mechanism for
'AddOrEdit' function in RecordExpensePage.xaml.cs. Now, AddOrEdit()
returns a tuple containing a success flag and an exception which
provides a more robust way of handling task completion. Further
functional improvements are still commented out and will be considered
in future commits.