ci(Update): Add expense recording prompt
Following the success of adding an expense to the history, a new feature
was implemented to prompt the user if they want to record a new expense.
If the user responds positively, the properties of the last recorded
expense are copied into the new expense to make the recording process
easier and faster.