ci(Update): Add 'Delete' functionality to Dashboard page
A 'Delete' menu item has been added to the context menu of the Dashboard
page. In both the English and French localization files, the phrase
'Delete this record' has been included as well. The associated click
event handler, 'MenuItemDeleteRecord_OnClick', has been drafted but is
yet to be fully implemented.