ci(Update): Refactor navigation and remove unnecessary Dashboard references
The commit simplifies navigation between the Dashboard and RecordExpense
pages by utilizing the 'typeof' operator. This also allows to remove the
direct reference of DashboardPage inside RecordExpensePage, as the
navigation and operations are now independent. All the unnecessary calls
to refresh methods on DashboardPage inside RecordExpensePage have been
removed as they're no longer needed with this new way of managing pages.