ci(Update): Add button event handlers and UI elements to WelcomePage
Added two new button event handlers, ButtonAddDataBase_OnClick and
ButtonDatabase_OnClick, to the WelcomePage.xaml.cs file. These handlers
currently throw NotImplementedExceptions. In the WelcomePage.xaml file,
several UI modifications were made: a StackPanel was added with a button
to trigger the AddDatabase event, and an ItemsControl was populated with
existing databases that each have a button to trigger the Database
event.