ci(Update): Add RemoveDatabaseFile window functionality
Implemented the functionality for the RemoveDatabaseFile window in the
MyExpenses application. This includes the creation of a new XAML file
for the window design, as well as implementing the functionality to add
and remove database files through button clicks. Adjustments have also
been made to the WelcomePage to incorporate the new window.