ci(Update): Add application icon to various windows
The application icon has been added to the title bar of multiple windows
throughout the application. The path to the icon was added in the XAML
of each window. This update improves visual consistency and brand
recognition by displaying a unified icon across all windows. Some unused
window settings related to the icon in the left-hand-side window
commands were removed as well.