ci(Update): Update database addition functionality
Modified the ButtonAddDataBase_OnClick method in WelcomePage.xaml.cs to
use a dialog window for database file name input. This update ensures
the user inputs the name for the new database file, and the selected
file name will be used instead of the previously hard-coded
"test.sqlite". RefreshExistingDatabases() is now called to reflect the
addition of the new database.