ci(Update): Add utility to set default database values
This commit introduces a utility for setting default values in
DataBaseContext. It adds a new file: DbContextHelperSetDefaultValues,
which includes methods to create default values, specifically for Known
Colors. Modifications have also been made in WelcomePage.xaml.cs to
utilize this new utility, ensuring that the default values are set every
time a database is copied.