ci(Update): Add "isOpen" checkbox to AddEditLocationWindow and update location features
This commit includes changing the AddEditLocationWindow’s textbox into a
Grid that contains both the original textbox and a new checkbox
("isOpen"). The mapping feature is also updated to reflect the same
changes, now revealing whether a place is open or not. The SQL tables,
and the various related classes and methods have been updated to support
this new feature. Also, a third blue marker style is added for Mapsui.