ci(Update): Add new DashBoardPage feature
A new feature, DashBoardPage, has been added to the Wpf Pages. This
includes an xaml and its code-behind file. The DashBoardPage feature
incorporates components like buttons, DataGrids, ComboBoxes, and a
PieChart within a Grid. Primarily, it provides operations for fetching
and displaying specific data from the database onto the dashboard. It
also provides operations for navigating to other pages.