ci(Update): Update test package versions
Updated the versions of the test packages used in the project. The
Microsoft.NET.Test.Sdk package was updated to version 17.10.0 and the
xunit.runner.visualstudio package was updated to version 2.8.1 across
all test projects. These updates ensure that we are using the most
current versions of these libraries for our tests.