ci(Update): Remove color enumeration code in MainWindow
The initialization code that created a list of color codes and saved it
to the database in MainWindow.xaml.cs was removed. This code was
unnecessary and was likely for testing purposes, as it created a
black-listed color enumeration and saved each color to a new TColor
object in the database.